dsp API updated

+++ packages/utils/c/include/utils.h
@@ -40,8 +40,7 @@
 #define FTS_RANDOM_RANGE (2147483648.)
 
 
-UTILS_API void
-utils_config(void);
+UTILS_API void utils_config(void);
 
 
 /***************************************************************************************
@@ -51,16 +50,12 @@ utils_config(void);
  */
 
 /* 31 bit random number generator */
-UTILS_API void 
-fts_random_set_seed(unsigned long ul);
+UTILS_API void fts_random_set_seed(unsigned int ul);
 
-UTILS_API unsigned long 
-fts_random(void);
+UTILS_API unsigned int fts_random(void);
 
 /* float random between -1. and 1. */
-UTILS_API float 
-fts_random_float(void);
-
+UTILS_API float fts_random_float(void);
 
 #define fts_random_range(min, max) ((min) + (((max) - (min)) * fts_random() / (double)FTS_RANDOM_MAX))
